Move form dialog out of Import Data Page
